Chore: Refactoring datasets.py step 2a - fixations tests in own file
